+++ pex/pip/tool.py
@@ -21,6 +21,7 @@ from pex.common import safe_mkdir, safe_mkdtemp
 from pex.compatibility import get_stderr_bytes_buffer, shlex_quote, urlparse
 from pex.dependency_configuration import DependencyConfiguration
 from pex.dist_metadata import Requirement
+from pex.fetcher import URLFetcher
 from pex.interpreter import PythonInterpreter
 from pex.jobs import Job
 from pex.network_configuration import NetworkConfiguration
@@ -53,6 +54,7 @@ if TYPE_CHECKING:
         Match,
         Optional,
         Sequence,
+        Text,
         Tuple,
         Union,
     )
@@ -64,8 +66,8 @@ else:
 
 @attr.s(frozen=True)
 class PipArgs(object):
-    indexes = attr.ib(default=None)  # type: Optional[Sequence[str]]
-    find_links = attr.ib(default=None)  # type: Optional[Iterable[str]]
+    indexes = attr.ib(default=None)  # type: Optional[Sequence[Text]]
+    find_links = attr.ib(default=None)  # type: Optional[Iterable[Text]]
     network_configuration = attr.ib(default=None)  # type: Optional[NetworkConfiguration]
 
     def iter(self, version):
@@ -217,9 +219,17 @@ class PackageIndexConfiguration(object):
         self,
         pip_version,  # type: PipVersionValue
         target,  # type:  Union[UniversalTarget, MarkerEnvironment]
+        requirement_files=None,  # type: Optional[Iterable[str]]
     ):
         # type: (...) -> Optional[DownloadObserver]
-        return package_repositories.patch(self.repos_configuration, pip_version, target)
+        return package_repositories.patch(
+            repos_configuration=self.repos_configuration.with_contained_repos(
+                requirement_files,
+                fetcher=URLFetcher(network_configuration=self.network_configuration),
+            ),
+            pip_version=pip_version,
+            target=target,
+        )
 
 
 if TYPE_CHECKING:
@@ -640,7 +650,9 @@ class Pip(object):
             ),
             (
                 package_index_configuration.patch(
-                    pip_version=self.version, target=universal_target or target.marker_environment
+                    pip_version=self.version,
+                    target=universal_target or target.marker_environment,
+                    requirement_files=requirement_files,
                 )
                 if package_index_configuration
                 else None