+++ pex/resolve/package_repository.py
@@ -6,12 +6,23 @@ from __future__ import absolute_import
 import itertools
 import os
 import re
+from collections import OrderedDict, defaultdict
 
 from pex.auth import PasswordEntry
 from pex.compatibility import string
 from pex.dist_metadata import Requirement, RequirementParseError
 from pex.exceptions import production_assert, reportable_unexpected_error_msg
+from pex.fetcher import URLFetcher
+from pex.orderedset import OrderedSet
 from pex.pep_503 import ProjectName
+from pex.requirements import (
+    FindLinks,
+    Index,
+    PyPIRequirement,
+    URLRequirement,
+    VCSRequirement,
+    parse_requirement_file,
+)
 from pex.resolve.target_system import MarkerEnv
 from pex.third_party.packaging.markers import InvalidMarker, Marker
 from pex.typing import TYPE_CHECKING, cast
@@ -19,7 +30,7 @@ from pex.typing import TYPE_CHECKING, cast
 if TYPE_CHECKING:
     # N.B.: The `re.Pattern` type is not available in all Python versions Pex supports.
     from re import Pattern  # type: ignore[attr-defined]
-    from typing import Any, Dict, Iterable, List, Optional, Tuple, Union
+    from typing import Any, DefaultDict, Dict, Iterable, List, Mapping, Optional, Text, Tuple, Union
 
     import attr  # vendor:skip
 
@@ -183,7 +194,7 @@ class Repo(object):
             location=data["location"], scopes=tuple(Scope.parse(scope) for scope in data["scopes"])
         )
 
-    location = attr.ib()  # type: str
+    location = attr.ib()  # type: Text
     scopes = attr.ib(default=())  # type: Tuple[Scope, ...]
 
     def as_dict(self):
@@ -240,8 +251,8 @@ class PackageRepositories(object):
     _target_env = attr.ib()  # type: Union[Dict[str, str], MarkerEnv]
     _scoped_indexes = attr.ib(default=())  # type: Tuple[Repo, ...]
     _scoped_find_links = attr.ib(default=())  # type: Tuple[Repo, ...]
-    global_indexes = attr.ib(default=(Repo(PYPI),))  # type: Tuple[str, ...]
-    global_find_links = attr.ib(default=())  # type: Tuple[str, ...]
+    global_indexes = attr.ib(default=(Repo(PYPI),))  # type: Tuple[Text, ...]
+    global_find_links = attr.ib(default=())  # type: Tuple[Text, ...]
 
     @property
     def has_scoped_repositories(self):
@@ -266,7 +277,7 @@ class PackageRepositories(object):
         scoped_repos,  # type: Iterable[Repo]
         project_name,  # type: ProjectName
     ):
-        # type: (...) -> List[str]
+        # type: (...) -> List[Text]
         return [
             repo.location
             for repo in scoped_repos
@@ -274,11 +285,11 @@ class PackageRepositories(object):
         ]
 
     def in_scope_indexes(self, project_name):
-        # type: (ProjectName) -> List[str]
+        # type: (ProjectName) -> List[Text]
         return self._in_scope_repos(scoped_repos=self._scoped_indexes, project_name=project_name)
 
     def in_scope_find_links(self, project_name):
-        # type: (ProjectName) -> List[str]
+        # type: (ProjectName) -> List[Text]
         return self._in_scope_repos(scoped_repos=self._scoped_find_links, project_name=project_name)
 
 
@@ -289,6 +300,7 @@ class ReposConfiguration(object):
         cls,
         indexes=(),  # type: Iterable[Repo]
         find_links=(),  # type: Iterable[Repo]
+        derive_scopes_from_requirements_files=False,  # type: bool
     ):
         # type: (...) -> ReposConfiguration
         password_entries = []
@@ -301,20 +313,78 @@ class ReposConfiguration(object):
             index_repos=tuple(indexes),
             find_links_repos=tuple(find_links),
             password_entries=tuple(password_entries),
+            derive_scopes_from_requirements_files=derive_scopes_from_requirements_files,
         )
 
     index_repos = attr.ib(default=(Repo(PYPI),))  # type: Tuple[Repo, ...]
     find_links_repos = attr.ib(default=())  # type: Tuple[Repo, ...]
     password_entries = attr.ib(default=())  # type: Tuple[PasswordEntry, ...]
+    derive_scopes_from_requirements_files = attr.ib(default=False)  # type: bool
+
+    def with_contained_repos(
+        self,
+        requirement_files=None,  # type: Optional[Iterable[Text]]
+        fetcher=None,  # type: Optional[URLFetcher]
+    ):
+        # type: (...) -> ReposConfiguration
+
+        if not requirement_files:
+            return self
+
+        indexes_by_source = OrderedDict()  # type: OrderedDict[Text, OrderedSet[Text]]
+        find_links_by_source = OrderedDict()  # type: OrderedDict[Text, OrderedSet[Text]]
+        scopes_by_source = defaultdict(OrderedSet)  # type: DefaultDict[Text, OrderedSet[Scope]]
+        for item in itertools.chain.from_iterable(
+            parse_requirement_file(requirement_file, fetcher=fetcher)
+            for requirement_file in requirement_files
+        ):
+            if self.derive_scopes_from_requirements_files and isinstance(
+                item, (PyPIRequirement, URLRequirement, VCSRequirement)
+            ):
+                scopes_by_source[item.line.source].add(
+                    Scope(project=item.requirement.project_name, marker=item.requirement.marker)
+                )
+            elif isinstance(item, FindLinks):
+                find_links_by_source.setdefault(item.line.source, OrderedSet()).add(item.location)
+            elif isinstance(item, Index):
+                indexes_by_source.setdefault(item.line.source, OrderedSet()).add(item.location)
+
+        if not indexes_by_source and not find_links_by_source:
+            return self
+
+        def merge_scopes(
+            repos,  # type: Iterable[Repo]
+            locations_by_source,  # type: Mapping[Text, Iterable[Text]]
+        ):
+            scopes_by_location = OrderedDict(
+                (repo.location, OrderedSet(repo.scopes)) for repo in repos
+            )
+            for source, locations in locations_by_source.items():
+                for location in locations:
+                    scopes_by_location.setdefault(location, OrderedSet()).update(
+                        scopes_by_source[source]
+                    )
+            return tuple(
+                Repo(location=location, scopes=tuple(scopes))
+                for location, scopes in scopes_by_location.items()
+            )
+
+        return attr.evolve(
+            self,
+            index_repos=merge_scopes(repos=self.index_repos, locations_by_source=indexes_by_source),
+            find_links_repos=merge_scopes(
+                repos=self.find_links_repos, locations_by_source=find_links_by_source
+            ),
+        )
 
     @property
     def indexes(self):
-        # type: () -> Tuple[str, ...]
+        # type: () -> Tuple[Text, ...]
         return tuple(repo.location for repo in self.index_repos if not repo.scopes)
 
     @property
     def find_links(self):
-        # type: () -> Tuple[str, ...]
+        # type: () -> Tuple[Text, ...]
         return tuple(repo.location for repo in self.find_links_repos if not repo.scopes)
 
     def scoped(self, target_env):